Factor out crate translation in tests
